ci: Use special pretty output format for tests in GitHub Actions.
Ginko supports --github-output flag that produces specially formatted
output for GHA logs.  Current kind script doesn't log anything until
the very end that makes it hard to debug timeouts and the control plane
script is using verbose output that is producing a lot of logs that
are hard to navigate.  Special format will provide the output and hide
unnecessary parts in expandable blocks.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>